Runs a Flask server to serve generated PDFs, adds Dockerfile
The Flask server currently takes no parameters, but it is planned that
instead of test data, it'll be a POST-only endpoint which takes a JSON
formatted payload and generates a PDF from it.

Additionally, the PDF is no longer stored to disk, and only exists in
memory and served out to the requester.